requirements.txt
deepspeed
ffmpeg
more_itertools
numpy
pandas
pytorch_lightning
PyYAML
regex
sacrebleu
torch
tqdm
transformers
pre-commit
datasets
soundfile
librosa
fairseq2
silero-vad
openai-whisper
speechtokenizer